Reorder actions in the parser
